backend/requirements.txt
Django
djangorestframework
django-cors-headers
psycopg2-binary
python-decouple
django-environ
Pillow
django-filter
pgvector
djangorestframework-simplejwt
celery
redis
PyPDF2
python-docx
python-magic-bin